Cleaned produced html source
Html elements that only contains a single html text will now be displayed on a mono line. For example :
<label for="plop">hello</label>.
Previously, generated source code would be :
<label for="plop">
	hello
</label>

This allows lighter source code.

to avoid a complete change of architecture (for an admitedly minor and not really mandatory change) this modification use a smal hack, and involved creating methods that shouldn't exist.
If any bugs are created because of this, Frederic will offer 3 candies per bug fix on this matter.