Adaptation of HtmlList to the new framework
HtmlList was part of an ancient API. It is now changed to be part of the new beautiful HtmlElement API.
You can now create simple HtmlList, numbered or not, without having to pass 1 billion useless parameters.
You can add list elements with a simple line, based on a simple string, or can give more complicated elements if needed.